+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added statistics about not-yet-completed Task Assignments to
+  Project Stats page.
 
 ## [2.3.0] - 2021-02-15
 ### Added